# src/main/resources/api-contract.yaml
openapi: 3.0.3
info:
  title: Customers and Products API
  version: 1.1.0
  description: API to manage customers and products for practice purposes
servers:
  - url: /system/api/v1
    description: Local server
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
          format: double
        stock:
          type: integer
      required:
        - id
        - name
        - price
        - stock
      example:
        id: 1
        name: "Monitor"
        price: 299.99
        stock: 5
...
paths:
...
  /products:
    get:
      summary: Get all products
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      summary: Add a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product added
          content:
            application/json:
              schema:
                type: string
                example: Product Ratón inalámbrico was created.
        '409':
          description: Product with that id or name already exists
          content:
            application/json:
              schema:
                type: string
                example: The product with name Monitor already exists.
  /products/filter:
    get:
      summary: Get all products between minPrice and maxPrice
      parameters:
        - name: minPrice
          in: query
          required: true
          schema:
            type: number
            example: 100.0
        - name: maxPrice
          in: query
          required: true
          schema:
            type: number
            example: 350.0
      responses:
        '200':
          description: Products between minPrice and maxPrice
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          description: Wrong request
          content:
            application/json:
              schema:
                type: string
                example: minPrice must exist
  /products/{id}/stock:
    patch:
      summary: Update the product's stock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stock:
                  type: integer
                  example: 12
      responses:
        '200':
          description: Stock updated correctly
          content:
            application/json:
              schema:
                type: string
                example: Stock for product id = 3 updated to 12
        '400':
          description: Number stock is null or negative
          content:
            application/json:
              schema:
                type: string
                example: Stock can't be negative
        '404':
          description: The product doesn't exist
          content:
            application/json:
              schema:
                type: string
                example: The product with id = 8 wasn't found
...